move GrpcService to its own file
